requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.0
jieba==0.39
lxml==4.8.0
numpy==1.21.6
pandas==1.0.5
PyMySQL==1.0.2
requests==2.27.1
wordcloud==1.8.2.2
snownlp==0.12.3

beautifulsoup4~=4.7.1
matplotlib~=3.3.0
Pillow~=7.2.0